refactor: Domain 과 UI로직 분리에 따른 파일 분리
