+++ src/UI/components/Header.js
@@ -1,5 +1,5 @@
-import { DOM } from '../dom.js';
-import { BUTTON_IMAGE_SRC } from '../../public/assets/imgaePaths.js';
+import { DOM } from '../../dom.js';
+import { BUTTON_IMAGE_SRC } from '../../../public/assets/imgaePaths.js';
 
 class Header {
   #addRestaurantModal;
