+++ src/UI/components/Modal.js
@@ -1,4 +1,4 @@
-import { DOM } from '../dom.js';
+import { DOM } from '../../dom.js';
 
 class Modal {
   #modal;
